raksha-backend/requirements.txt
# Web Framework & Server
fastapi
uvicorn[standard]
websockets

# Database (Supabase)
supabase

# AWS & AI Integration
google-generativeai
httpx

# Environment & Data Validation
python-dotenv
pydantic
pydantic-settings
python-multipart